update kubecfg to latest
Kubecfg contains an improvement to the sort utility which, in
particular, ensures CustomResourceDefinitions are created before CRD
resources themselves. This is required for the upcoming AppRepository
stuff (improving the chart repo syncing).

I've updated Kubecfg to the latest commit as not much else has changed
since the sort improvement and latest.