chart-repo: fix bug indexing with missing files
Previously, when attempting to extract multiple files (README and
values) from the tarball, we were not resetting the file pointer when
reading through the tarball. This meant that only the first file we
searched for would actually be picked up.

This changes the logic to search for all files at the same time when
passing through the tarball, without needing to rewind the file pointer.