Use static nginx frontend for Kubeapps services
- Removes Kubeapps' internal nginx ingress controller in favour of a
  static nginx deployment and config

We don't really need the Ingress controller, and it tends to confuse
users as they usually want to use their own Ingress controllers. With
this change, users can treat Kubeapps as a single service (the
`kubeapps` service) and setup their own Ingress controllers to front it.

- Removes kubernetes API proxy pod in favour of routing directly to the
  `kubernetes.default` service

This ensures we always use the credentials Kubeapps is logged in with
instead of a service account, and also removes an extra pod.