manifests: Use non-persistent emptyDir volume for MongoDB
Kubeapps uses mongodb for cache storage, we do not really need to attach persistent storage for this cache as it can be easily be populated automatically or on demand. This commit swaps out the pvc volume with a emptyDir volume which should not only reduce the system requirements of running kubeapps but also help alleviate any mongodb file permission issues.

closes #336