ignore no charts error when filtering
Previously, we would display the No Charts Found error alert if the
filtered chart array came back empty. However, this error only makes
sense to show when the charts from the server come back empty. This
patch fixes this to only show the error if there are no charts from
the server.

We may want to improve the way we show this message (for example, under
the Charts header), and we may also want to show a different alert if
the filter shows up empty in the future.