only update listOverview if release present when processing updateInfo
We weren't checking if we were actually able to find the release name in
the `listOverview`, if it wasn't there (which is usually the case after
installing a chart as you get directed to the app page not the app list)
then `appOverviewIndex` would be `-1` and we won't be slicing the array
correctly.

This patch checks that we actually found the release in the
`listOverview` before updating the entry. If we don't find it, we don't
need to do anything, as it will get fetched and set in the
`listOverview` when visiting the AppList page.