Replace Object.keys with Object.entries
This is necessary for the `component` React property to be properly type
checked. The reason is that, as far as TypeScript knows, `Object.keys`
only returns plain strings, whereas `component` in `[route, component]`
is assigned as narrow a type as expected, derived from the values in the
routes object.

Another advantage is that the routes objects are not repeated inside the
mapped functions, so there is no risk of writing e.g.

    <Route ... component={privateRoutes[route]} />

instead of

    <Route ... component={routes[route]} />