Remove post-install annotation for apprepositories (#1754)
We had added this to ensure that during installation of Kubeapps, the
creation of the app repository resources wouldn't trigger a sync job
to run before the upgrade job updates the schema. But one case we did
not anticipate is that if the installation is held up (network issues
or similar) and the helm install fails, the resources won't be present.

I think it's better to ensure they are always present so that the sync
job can eventually sync them (and update to safely migrate the schema
rather than recreating it - as per #1560).