Support charts with a '/' in their names (#2173)
* Delete --database-type in devel docs since it is no longer needed

* Add port in default database-url

* Allow chart names with '/' in backend asset-syncer

* Allow chart names with '/' in backend nginx

It passes downstream the encoded URI, without a prior decoding. This involves changes in downstream services.
At the moment, only "/api/assetsvc" is being modified

* Allow chart names with '/' in backend assetsvc

Since it is possible to receive unescaped parameters (like '%2F' --> '/') , this commit adds support for manually decoding each param and handle it accordingly.

* Allow chart names with '/' in frontend

* Allow chart names with '/' in backend assetsvc

More changes here

* Fix icon url in chart catalog item

A double // was being added.
Generated in assetsvc/handler.go:337

* Make uniqueChartList function repo aware

A chart is skipped if is a duplicated in the same repo

* Fix icon url in chart catalog item (deploy view)

A double // was being added.
Generated in assetsvc/handler.go:337

* Make getPaginatedChartList duplicate filtering repo-aware

A chart is skipped if is a duplicated in the same repo, not in the whole list.

* Fix deployment info url after the enconding change

* Improve chart icon url

Some test were assuming the chartAttrs.icon to contain a string not starting with '/'.
This commit is to support both alternatives

* Filter out duplicate repos in upgrade form

* Fix update info url after the enconding change

Now nginx will concat (and duplicate) query params to ensure they arrive properly encoded

* Minor fixes after a not passing ci

* Add non-strict failback chart search in db queries

When trying to upgrade an app with slashes in its name (since it is stored with '/' in db) no chart was being found. With this non strict "repo+whatever+charName" fallback search the chances of obtaining empty results decrease.

* Fix unnoticed wrong committed file

A commented version of the test was committed by mistake

* Remove unnoticed comment

* Use the same icon URL logic as in chartheader

* Revert uniqChartList changes

Now it is unscoped again

* Replaced getNameFromLastSlash to path.Base

It's important to learn the default libraries before implementing yourself...

* Revert fetchRepos changes

* Simplify icon url logic

* Update chart assestsvc url

* Improve nginx assetsvc matching

Now it matches from "/api/assetsvc/what$2Fever?param=value" as $1="/what$2Fever" and $args="param=value"
So, downstream services will receive "/assetsvc/what$2Fever?param=value"

* Add additional test for the 'with slashes' case instead of replacing it

* Improve error handling when decoding url params

* Remove unnoticed committed lines

* Store encoded chartName into DB.

* Display properly encoded chartNames in UI

* Send decoded chartName in http body

* Assetsvc no longer decode/encode chartNames. Update tests accordingly

* Remove redundant nginx rewrite

* Revert useless var rename

* Use chartID instead of the aux function

* Revert change in dups filtering

It's already part of another PR

* Rename misleading var name

Now it doesn't make sense to distinguish between escaped and unescaped there

* Fix failing outdated test

* Minor fixes in code comments