Create correct secrets for namespaced repos with authorization
For namespaced repos we need to create two identical secrets containing
ca.crt and/or authorizationHeader. We need one in the namespace that the
app repository is scoped to and one in the kubeapps release namespace.

The name of the secret in the kubeapps release namespace is expected to
be of the format {{ .namespace }}-apprepo-{{ .name }} by the
apprepository controller.