Add ability to set annotations to each individual deployment
The reason I want this supported is so that I can add these deployments
to my service mesh. Most service mesh decide on which deployments/pods
should be part of the mesh using annotations. I would like to be able to
control this via the Helm chart.

Service meshes can also allow you to add a whole namespace to the
service mesh. The reason this doesn't work well is that the sync jobs
created by kubeapps will never complete if they have a sidecar
container, which is what's done to add a pod to a service mesh.

This only adds support for adding annotations to the deployments, not
the pods created by those deployments. If we want to support that as
well, we could add another field called podAnnotations to each of these
components.