Update so plugins are passed a rest.Config getter rather than a client getter (#3161)
* Update so plugins are passed a rest.Config getter rather than a client getter

Originally we passed plugins a client getter that return a dynamic
client. We then found that the helm plugin additionally needs a typed
client, so updated the client getter to return both. For work that I'm
currently doing instantiating a helm API action configuration, I also
need the plugin to have access to a genericclioptions.RESTClientGetter.

Rather than updating the existing client getter to return yet another
type, and given that it's a single line to create a client from the rest
config, I'm instead updating so that plugins are initialised with an
initialized rest.Config getter so that they can create any client they
need while still not needing to know any details of the config (eg.
pinniped setup or similar).

Signed-off-by: Michael Nelson <minelson@vmware.com>

* Rename a couple of clientGetter vars to configGetter.

Signed-off-by: Michael Nelson <minelson@vmware.com>