Avoid docker production images overlapping between GHA and CircleCI (#5528)
So far, when Docker production images are generated from the GHA
kubeapps general workflow, those are pushed to Docker Hub in the
`kubeapps` workspace (eg. `kubeapps/kubeapps-apis`,
`kubeapps/dashboard`, etc) and tagged with `latest` or the release
version `x.y.z`, so they overlap with the ones generated in the CircleCI
workflow that runs in parallel (until we finish the migration to GHA and
decommission the CircleCI workflow). To avoid images from both workflows
to overwrite each other, this PR adds a `-gha` suffix to the tag for the
images generated from the GHA workflow (eg. `kubeapps/dashboard:latest-gha)`.

Also, I took advantage of this PR to fix the step `Show PR Context` from
the GHA kubeapps general workflow's `setup` job. This step is mainly for
debugging purposes.