Consistent json schema validation of SliderParams (#5533)
### Description of the change

As pointed out by @dud225, we weren't enforcing the numeric validation
consistently in both the slider and the text input. Besides, we were
missing some json schema fields like `exclusiveMinimum` and
`exclusiveMaximum`.
This PR is to fix that.

### Benefits

Consistent param validation

### Possible drawbacks

N/A

### Applicable issues

- fixes #5520

### Additional information

> **Note**
> This PR is part of a series of PRs aimed at closing [this
milestone](https://github.com/vmware-tanzu/kubeapps/milestone/27). I
have split the changes to ease the review process, but as there are many
interrelated changes, the tests will be performed in a separate PR (on
top of the branch containing all the changes).
>  PR 4 out of 6



````json
    "number-maxmin": {
      "type": "number",
      "maximum": 10,
      "exclusiveMaximum": 9,
      "minimum": 2,
      "exclusiveMinimum": 1
    },
`````


![image](https://user-images.githubusercontent.com/11535726/197026623-2c2a5284-f939-4af2-a1a2-9816cf65234d.png)


![image](https://user-images.githubusercontent.com/11535726/197026701-370c21d3-7164-405d-b77b-56dd58472dbb.png)


Also, note the HTML5 additional validation (user's browser locale msgs):


![image](https://user-images.githubusercontent.com/11535726/197026840-b5dd2e98-245b-4b55-b52a-1deaf0746866.png)

Signed-off-by: Antonio Gamez Diaz <agamez@vmware.com>