Add JSON Schema editor view (#5530)
### Description of the change

The basic forms view has been recently revamped with plain JSON schema
support. However, most of the charts (and maybe other packages) do not
bring any schema file as part of their bundle.

#3535 brings some ideas on how to solve it, but, this PR is just paving
the way with a very simplistic approach: let users edit/add a schema
from a text editor.

Currently, there is no way to upload a file externally, but if this
happens to be helpful, could be added in a follow-up PR.

### Benefits

Users will be able to edit/add json schemas for improving the "basic
form" experience (validations, defaults, etc...).

### Possible drawbacks

Important: this "edited (or added)" schema is NOT being used on the
backend side. Meaning you can't expect a proper validation or any
further schema enforcement. Using the schema in the backend as well
would imply some extra changes that could be performed in a follow-up PR
if really needed.

### Applicable issues

- fixes #3535

### Additional information

> **Note**
> This PR is part of a series of PRs aimed at closing [this
milestone](https://github.com/vmware-tanzu/kubeapps/milestone/27). I
have split the changes to ease the review process, but as there are many
interrelated changes, the tests will be performed in a separate PR (on
top of the branch containing all the changes).
>  PR 1 out of 6


![PR-1](https://user-images.githubusercontent.com/11535726/197021211-e64814a7-29b1-4ac7-8043-51075f9dda9d.gif)

Signed-off-by: Antonio Gamez Diaz <agamez@vmware.com>