Migrate prerelease flow to GHA (#5590)
This PR turns the current `Kubeapps general pipeline` workflow into a
general reusable workflow, and adds two new workflows that call the former:
* Full Integration Pipeline: it is a new workflow that is manually
triggered from the GitHub Actions dashboard. It calls the general
workflow providing the input `run_gke_tests` with the value `true`, so
the set of e2e tests is executed on the GKE environments.
* Main Pipeline: it is a workflow triggered automatically on `push`
and `pull request` events. It calls the general workflow without
providing the input parameter `run_gke_tests` (its default value is
`false`), so the GKE tests aren't run. This workflow comes to replace
the current `kubeapps general pipeline` workflow that has been turned
into a reusable workflow, so we need this one to trigger it
automatically.

Signed-off-by: Jesús Benito Calzada <bjesus@vmware.com>