Remove "isLoaded" flag from deployment form (#5605)
### Description of the change

This PR solves a bug found during a manual test prior to the release:
when selecting a different version, the `isLoaded` flag was preventing a
proper execution of the useEffect flow. The flag has been replaced with
`isLoading`, which just avoid re-renders when an expensive operation is
running (like parsing the schema).

### Benefits

Users upgrading package will see the deployed values (correct) instead
of the pkg defaults (incorrect).

### Possible drawbacks

In those cases, the component will re-render twice, but that's not a big
deal (no expensive backend calls, the table is now performing well,
etc).

### Applicable issues

- fixes #5603

### Additional information


![deployedValuesFix](https://user-images.githubusercontent.com/11535726/199706432-720c65c0-8351-438d-86d0-bbf3192a1955.gif)

Signed-off-by: Antonio Gamez Diaz <agamez@vmware.com>