Run GHA's Full Integration Pipeline on schedule (#5621)
With this commit, GitHub Action's `Full Integration Pipeline` is run on schedule. We usually 
run this pipeline manually only when we are going to release a new version of Kubeapps, 
and from time to time it turns out that it fails. With this change, we get a much shorter 
feedback loop, so we can act on the errors more proactively.

Signed-off-by: Jesús Benito Calzada <bjesus@vmware.com>