Extract codeql from GHA linters (#5629)
In this PR we take the `CodeQL` linter out of the GHA `linters.yml`
workflow, so it is no longer part of the `Kubeapps Main Pipeline`. We
have taken this decision because the CodeQL analysis for `golang` is
really slow (~30 min) and delays the pipelines, while the value it adds
is not so clear. That being said, we have extracted to an independent
workflow as it was before introducing the reusable linters workflow.

Signed-off-by: Jesús Benito Calzada <bjesus@vmware.com>