Add status check jobs for local_e2e_tests matrix and linters (#5628)
Due to GH limitations, it is not possible to add a branch protection
rule pointing to all the jobs generated by GHA's matrix strategy. As a
workaround, we've introduced a new status check job that depends on the
job with the matrix, and whose status depends on the success of all the
jobs generated by that. In addition, we've taken advantage to add
another check-status job for all the linter jobs.

Signed-off-by: Jesús Miguel Benito Calzada <bjesus@vmware.com>