Re-enable hash tests after unsuccessfully trying to repro. (#5550)
Signed-off-by: Michael Nelson <minelson@vmware.com>

### Description of the change

<!-- Describe the scope of your change - i.e. what the change does. -->
I'd created #4900 in June after noticing occasional failures, but then
not seeing those in CI at all, so we didn't land the PR until Antonio
noted the errors occurring again nearly 3 months later.

The issue was that part of the `TokenCredentialRequest` that is
generated for the test depends on an env variable (specifically,
`DEFAULT_PINNIPED_API_SUFFIX`) which was also set in another test
`test_get_api_group_getters`. Rust runs tests concurrently, so if the
test happened to be run in a thread of the same process *before* the
hash test, then they would generate a different (but predictable) hash.
The actual different hash can be reproduced simply by setting the env
var manually to the same value set in the test.

The solution is to ensure that the tests which modify the environment
restore the environment after completing, and that any test which
depends on those environment variables run with a mutex (so they don't
run concurrently). Hence using the `temp-env` crate which can be used to
ensure just that.

### Benefits

No more transient failures for these tests.

### Possible drawbacks

None

### Applicable issues

<!-- Enter any applicable Issues here (You can reference an issue using
#) -->

- Fixes #5296

### Additional information

<!-- If there's anything else that's important and relevant to your pull
request, mention that information here.-->

Signed-off-by: Michael Nelson <minelson@vmware.com>