Issue5542 - Unable to edit or delete newly created flux package repository (#5664)
### Description of the change

This change completes the changes previously made in an earlier change
regarding flux repository edit/delete buttons enablement issues.
The main switch is to change the scope of a flux repository from global
to namespaced. A flux repository behaves more similar to a namespaced
repository than a global repository, making its handling easier in the
UI.

The change also includes a fix where carvel global repositories were not
included, even if the user has list/view access to them.

Finally, the create/update repository dialog in the UI has been updated
to reflect that the default scope of a flux repository is namespaced,
and not global, and that it cannot be changed.

### Benefits

Users that have namespace-only access can now properly manage the flux
repositories they created.

### Applicable issues

- fixes #5542

### Additional information

This change does not include a fix for the page and table buttons when
the "show in all namespaces" toggle is on.
Users can use the workaround to toggle off the switch. For this reason,
a proper fix is best delayed to when the repository page is revamped.