Fix GHA release workflow (#5717)
This commit aims to fix the `Release Pipeline` of the GHA workflows
(better said, to make it support testing releases, those triggered by
tags following the pattern `test-vX.Y.Z`). Currently, it is failing
because in the `Kubeapps General` workflow, we are setting the
`running_on_tag` output variable of the `setup` job based on whether the
Git ref type is `tag` and it matches the pattern `vX.Y.Z`. Given that we
are controlling what tag patterns can trigger a release in the `Release
Pipeline` workflow, it doesn't make sense to have this condition, as we
need to make sure that it's in sync with what there is in the workflow
conditions, and it's error-prone.

Signed-off-by: Jesús Benito Calzada <bjesus@vmware.com>