Simplify release-related logic in GHA's kubeapps general pipeline (#5718)
This commit aims to simplify the logic in the GHA's Kubeapps General
workflow. Currently, we have some logic that checks if the pipeline has
been triggered by a tag pushed to the repo, and if it is, we run some
jobs: push_images, sync_charts, release, etc. As we already have a
top-level workflow `kubeapps-release.yml` that internally uses the
general workflow, and only runs when a tag following the release pattern
is pushed to the repo, we can just introduce an input parameter in the
general workflow that controls whether those jobs should be run or not,
instead of relying on the logic residing in the general pipeline.

Signed-off-by: Jesús Benito Calzada <bjesus@vmware.com>