Finish switching from GKE_BRANCH to GKE_VERSION (from #5730) (#5780)
This PR aims to fix the failing full integration pipeline (which we need
to pass for release).

Also updates Hugo version (based on instructions in the
release-process.md.

Signed-off-by: Michael Nelson <minelson@vmware.com>

<!--
Before you open the request please review the following guidelines and
tips to help it be more easily integrated:

 - Describe the scope of your change - i.e. what the change does.
 - Describe any known limitations with your change.
- Please run any tests or examples that can exercise your modified code.

 Thank you for contributing!
 -->

### Description of the change

<!-- Describe the scope of your change - i.e. what the change does. -->
Recent runs (from the past 6 days) of the [full integration
pipeline](https://github.com/vmware-tanzu/kubeapps/actions/workflows/kubeapps-full-integration.yaml)
have been failing. Looking for the differences, I could see that:

- `chart-museum.sh` was failing to upload our custom apache chart to the
chartmuseum instance, because...
- `LOAD_BALANCER_IP` was set to `172.18.0.2`, which is the `DEX_IP`,
whereas it should be set to the load balancer IP address (which has been
created correctly), as it is in the last passing run. It's not doing
this because...
- The code that sets the `LOAD_BALANCER_IP` address does so
conditionally based on the `GKE_BRANCH` environment variable, which is
*blank*... so as far as that script can tell, it's not a GKE run. The
reason that the `GKE_BRANCH` env var is blank is because...
- In
[#5730](https://github.com/vmware-tanzu/kubeapps/pull/5730/files#diff-164a87eb83a2101307c123bd6651d8f9355b51a995971d260038a1be44d6f6dbL19)
the `GKE_BRANCH` env var was removed and replaced with `GKE_VERSION`
everywhere *except* `e2e-test.sh` it seems. I assume this was just an
oversight on our part.

So this PR just finishes that change by renaming the var in
`e2e-test.sh`.

### Benefits

<!-- What benefits will be realized by the code change? -->

Hopefully we can now pass the full integration pipeline.

### Possible drawbacks

<!-- Describe any known limitations with your change -->

### Applicable issues

<!-- Enter any applicable Issues here (You can reference an issue using
#) -->

- fixes #

### Additional information

<!-- If there's anything else that's important and relevant to your pull
request, mention that information here.-->

Signed-off-by: Michael Nelson <minelson@vmware.com>