Add more helm custom fields (#5622)
### Description of the change

This PR brings more custom fields to the Helm plugin (repos API).
Specifically, I've mimicked the currently customizable values through
the `values.yaml` file. That is: `tolerations`, `nodeSelector`,
`securityContext` and `proxyOptions`.
Notwithstanding, users were able to set their custom `PodTemplateSpec`
for the sync job in the old fashion way, but there's too much
information to be included as a custom value in the repos API. But why?

Well, since we are generating code from the grpc messages, if we define
a `pod_template_spec` field of type `PodTemplateSpec` in the grpc proto
file.... we need to have the corresponding type defined.
In doing so, there's two alternatives: 1) using an existing definition
and importing it using Buf (via BSR) or 2) maintaining ourselves the
grpc messages definition.

- Option 1: almost not viable. There is no official definition in BSR,
just one coming from a private company with no guarantees whatsoever.
I've explored that way, but I don't feel we should go in this direction.
Too risky.
- Option 2: if we are to maintain the messages, we have two options:
2.1) using the one from Kubernetes or 2.2) defining our own model.
- Option 2.1: the autogenerated file is huge (=every k8s API object
model...) and we would have to ensure we're in sync when a new k8s
version is out.
- Option 2.2: it would require copying a subset of messages from the k8s
API, but keeping just those ones we want.
 
I've gone with option 2.2, but simplifying the fields we support. Just 1
or 2 objects having simple types like string or bool.
I'm adding some PR review messages to highlight what I mean.

That said, I'm open to any other better ideas you may have.


On the other hand, there's a separate issue: what should we display in
the Repos UI?
Well, in this case, I've opted for just showing the proxy options. The
rest of the fields are just related to the PodTemplateSpec and I feel
they are too complex for even advanced users. If someone really wants to
tune them up, they can just go an invoke the API.


### Benefits

Users will be able to customize the proxy options (along with a few
other options) via the repos API instead of just using the chart
`initialRepos` field.

### Possible drawbacks

Not every custom field is available through the UI, but I don't really
think we want to.

### Applicable issues

- fixes #5128

### Additional information

N/A

Signed-off-by: Antonio Gamez Diaz <agamez@vmware.com>
Signed-off-by: Antonio Gámez, PhD <antgamdia@gmail.com>
Signed-off-by: Michael Nelson <minelson@vmware.com>
Co-authored-by: Michael Nelson <absoludity@gmail.com>
Co-authored-by: Michael Nelson <minelson@vmware.com>